// packages/frontend/src/api/client/models/ProfileDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProfileDto = {
    /**
     * Profile numeric identifier
     */
    id: number;
    /**
     * Profile name
     */
    name: string;
    /**
     * Optional profile description
     */
    description: Record<string, any> | null;
    /**
     * Creation timestamp (ISO 8601)
     */
    created_at: string;
    /**
     * Last update timestamp (ISO 8601)
     */
    updated_at: string;
};