// packages/frontend/src/api/client/models/ProfileUpdateDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProfileUpdateDto = {
    /**
     * Profile name
     */
    name?: string;
    /**
     * Profile description
     */
    description?: Record<string, any>;
};